# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v2/tenants/{tenantId}/commands/batch": {
            "post": {
                "description": "向多个设备或分组发送批量命令",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "命令管理"
                ],
                "summary": "发送批量命令",
                "parameters": [
                    {
                        "type": "string",
                        "description": "租户ID",
                        "name": "tenantId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "批量命令请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.BatchCommandRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v2/tenants/{tenantId}/commands/{commandId}": {
            "get": {
                "description": "根据命令ID获取命令详细信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "命令管理"
                ],
                "summary": "获取命令详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "租户ID",
                        "name": "tenantId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "命令ID",
                        "name": "commandId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Command"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "取消指定的待执行命令",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "命令管理"
                ],
                "summary": "取消命令",
                "parameters": [
                    {
                        "type": "string",
                        "description": "租户ID",
                        "name": "tenantId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "命令ID",
                        "name": "commandId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v2/tenants/{tenantId}/devices/status": {
            "get": {
                "description": "获取指定租户下所有设备的状态概览",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "状态管理"
                ],
                "summary": "获取所有设备状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "租户ID",
                        "name": "tenantId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v2/tenants/{tenantId}/devices/{deviceId}/commands": {
            "post": {
                "description": "向指定设备发送控制命令（如屏幕开关、导航、音频等）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "命令管理"
                ],
                "summary": "发送设备命令",
                "parameters": [
                    {
                        "type": "string",
                        "description": "租户ID",
                        "name": "tenantId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "设备ID",
                        "name": "deviceId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "命令请求",
                        "name": "command",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CommandRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.CommandResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v2/tenants/{tenantId}/devices/{deviceId}/commands/history": {
            "get": {
                "description": "获取指定设备的命令历史记录",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "命令管理"
                ],
                "summary": "获取命令历史",
                "parameters": [
                    {
                        "type": "string",
                        "description": "租户ID",
                        "name": "tenantId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "设备ID",
                        "name": "deviceId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "每页数量",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "偏移量",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v2/tenants/{tenantId}/devices/{deviceId}/status": {
            "get": {
                "description": "获取指定设备的当前状态信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "状态管理"
                ],
                "summary": "获取设备状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "租户ID",
                        "name": "tenantId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "设备ID",
                        "name": "deviceId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DeviceStatus"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.BatchCommandRequest": {
            "type": "object",
            "properties": {
                "all": {
                    "type": "boolean"
                },
                "device_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "group_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "params": {
                    "type": "object",
                    "additionalProperties": true
                },
                "timeout": {
                    "type": "integer",
                    "example": 30
                },
                "type": {
                    "type": "string",
                    "example": "setScreen"
                }
            }
        },
...
        "api.CommandRequest": {
            "type": "object",
            "required": [
                "type"
            ],
            "properties": {
                "params": {
                    "type": "object",
                    "additionalProperties": true
                },
                "priority": {
                    "type": "integer",
                    "example": 0
                },
                "timeout": {
                    "type": "integer",
                    "example": 30
                },
                "type": {
                    "type": "string",
                    "example": "setScreen"
                }
            }
        },
        "api.CommandResponse": {
            "type": "object",
            "properties": {
                "command_id": {
                    "type": "string",
                    "example": "cmd-uuid-1234"
                },
                "duration": {
                    "type": "integer",
                    "example": 150
                },
                "error": {
                    "type": "string",
                    "example": ""
                },
                "result": {
                    "type": "object",
                    "additionalProperties": true
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "error message"
                }
            }
        },
...
        "models.Command": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "params": {
                    "type": "object",
                    "additionalProperties": true
                },
                "priority": {
                    "type": "integer"
                },
                "timeout": {
                    "description": "秒",
                    "type": "integer"
                },
                "timestamp": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/models.CommandType"
                }
            }
        },
...
        "models.CommandType": {
            "type": "string",
            "enum": [
                "setBrightness",
                "setScreen",
                "setVolume",
                "navigate",
                "reload",
                "reboot",
                "clearCache",
                "updateApp",
                "screenshot",
                "getLogs",
                "getWifiInfo",
                "installApp",
                "uninstallApp",
                "setRotation",
                "setKioskMode",
                "playSound",
                "stopSound",
                "speak",
                "wakeUp",
                "sleep",
                "setPin"
            ],
            "x-enum-varnames": [
                "CommandSetBrightness",
                "CommandSetScreen",
                "CommandSetVolume",
                "CommandNavigate",
                "CommandReload",
                "CommandReboot",
                "CommandClearCache",
                "CommandUpdateApp",
                "CommandScreenshot",
                "CommandGetLogs",
                "CommandGetWifiInfo",
                "CommandInstallApp",
                "CommandUninstallApp",
                "CommandSetRotation",
                "CommandSetKioskMode",
                "CommandPlaySound",
                "CommandStopSound",
                "CommandSpeak",
                "CommandWakeUp",
                "CommandSleep",
                "CommandSetPin"
            ]
        },
...
        "models.DeviceStatus": {
            "type": "string",
            "enum": [
                "pending",
                "active",
                "disabled",
                "deleted"
            ],
            "x-enum-varnames": [
                "DeviceStatusPending",
                "DeviceStatusActive",
                "DeviceStatusDisabled",
                "DeviceStatusDeleted"
            ]
        },
...
    }
}